fixed bug with edit posting
fixed a bug where editing a list item caused it to permanently disappear
from the list display page. added a boolean property is_complete so that
a single edit post method can bind a generic "list item" object and then
save it to the database as either "completed list item" or "uncompleted
list item" object based on the boolean value. added a checkbox to the
edit page so that user can change the is_complete property (thus
changing from the edit page whether item is displayed as complete or
uncomplete). also fixed bug where edit method did not preserve the
User_ID property of edited object (which also prevents display on the
user's to do list, causing item to disappear).